int->byte

+++ Mikroszimuláció/Mikroszimuláció/Form1.cs
@@ -119,7 +119,7 @@ namespace Mikroszimuláció
 
                     birthProbabilities.Add(new BirthProbability()
                     {
-                        Age = int.Parse(line[0]),
+                        Age = byte.Parse(line[0]),
                         NbrOfChildren = int.Parse(line[1]),
                         P = double.Parse(line[2])
                     });
@@ -141,7 +141,7 @@ namespace Mikroszimuláció
                     deathProbabilities.Add(new DeathProbability()
                     {
                         Gender = (Gender)Enum.Parse(typeof(Gender), line[0]),
-                        Age = int.Parse(line[1]),
+                        Age = byte.Parse(line[1]),
                         P = double.Parse(line[2])       //vesző pontra kicserélése, de én gépemen nem kellett .Replace(",",".")
                     });
                 }
